test: added a test for the rules mapping
