+++ sonar-erroraway-sonar-plugin/src/test/java/com/github/erroraway/sonarqube/ErrorAwayRulesDefinitionTest.java
@@ -15,7 +15,6 @@
  */
 package com.github.erroraway.sonarqube;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -27,11 +26,7 @@ import org.sonar.api.server.rule.RulesDefinition.Context;
 import org.sonar.api.server.rule.RulesDefinition.NewRepository;
 import org.sonar.api.server.rule.RulesDefinition.NewRule;
 
-import com.github.erroraway.ErrorAwayException;
 import com.github.erroraway.rules.ErrorAwayRulesMapping;
-import com.google.errorprone.BugPattern;
-import com.google.errorprone.BugPattern.SeverityLevel;
-import com.google.errorprone.bugpatterns.BugChecker;
 
 /**
  * @author Guillaume
@@ -64,16 +59,5 @@ class ErrorAwayRulesDefinitionTest {
 		verify(context, times(1)).createRepository(ErrorAwayRulesMapping.AUTODISPOSE2_REPOSITORY, "java");
 
 		verify(newRepository, times(ErrorAwayRulesMapping.RULES_COUNT)).createRule(Mockito.anyString());
-	}
-
-	@Test
-	void unknownRepository() {
-		assertThrows(ErrorAwayException.class, () -> ErrorAwayRulesMapping.repository(UnknownBugChecker.class));
-	}
-
-	@BugPattern(summary = "", severity = SeverityLevel.ERROR)
-	private static class UnknownBugChecker extends BugChecker {
-		private static final long serialVersionUID = 1L;
-
 	}
 }
